Move acronym logic into cog
