+++ src/ilo/jasima.py
@@ -1,5 +1,3 @@
-#!/bin/python3
-
 import json
 import urllib.request
